requirements.txt
flask
flask-sqlalchemy
flask-caching
python-dotenv
click
gunicorn
mysqlclient
redis
cryptography